demo/requirements/base.txt
django==1.11.7
psycopg2==2.7.3.2
redis==2.10.6
django-configurations==2.0
django-extensions==1.9.6
werkzeug==0.12.2
pyjwt==1.5.3
django-rediser